+++ here-naksha-lib-model/src/commonMain/kotlin/naksha/model/Tuple.kt
@@ -29,6 +29,7 @@ import naksha.base.NakshaError.NakshaErrorCompanion.ILLEGAL_STATE
 import naksha.base.NakshaException
 import naksha.base.Platform
 import naksha.base.Platform.PlatformCompanion.UNDEFINED
+import naksha.base.Platform.PlatformCompanion.logger
 import naksha.base.PlatformListApi.PlatformListApiCompanion.array_get
 import naksha.base.PlatformListApi.PlatformListApiCompanion.array_set
 import naksha.base.PlatformMapApi.PlatformMapApiCompanion.map_set
@@ -38,6 +39,8 @@ import naksha.base.Version
 import naksha.base.illegalArg
 import naksha.base.illegalState
 import naksha.base.internalError
+import naksha.base.proxy
+import naksha.geo.GeoUtil.GeoUtil_C.fromTWKB
 import naksha.model.objects.JsonPath
 import naksha.model.objects.MemberType
 import naksha.model.objects.NakshaCollection
@@ -63,21 +66,21 @@ import kotlin.jvm.JvmStatic
 @JsExport
 data class Tuple @JvmOverloads constructor(
     /**
-     * Feature serialized with the encoding described by the collection's dataEncoding.
+     * The uncompressed `JBON` binary of the Feature, to be read with `JbonDecoder`.
      * @since 3.0
      */
     @JvmField val featureBytes: ByteArray,
 
     /**
-     * The members book provided by storage at read time. Contains dedicated member values, such as `id`, `tn`, etc.
+     * The _HEAP_ decoded members book provided by storage at read time. Contains the values of the [Member].
      * @since 3.0
      */
     @JvmField val membersBook: IBook,
 
     /**
-     * After encoding a [NakshaFeature] into a [Tuple] using [encodeFeature] method, the [previousTupleNumber] will be set by the encoder to the [naksha.base.TupleNumber] of the given feature; if it had any.
+     * After encoding an object into a [Tuple] using [encodeFeature] method, the [previousTupleNumber] will be set in the generated [Tuple]. It will only be set by the encoder, if the given object contains a [TupleNumber][naksha.base.TupleNumber] of the state it had before being modified. In the XYZ specification this information was historically located in the jsonpath `properties->@ns:com:here:xyz->uuid`, but with modern `lib-base` support, the position can be freely selected.
      *
-     * This is metadata, it can as well be set manually, when a tuple is read from a storage.
+     * This is metadata, it can as well be set manually, when a tuple is read from a storage. It is not significant for the storage itself.
      * @since 3.0
      */
     @JvmField var previousTupleNumber: TupleNumber? = null
@@ -232,8 +235,8 @@ data class Tuple @JvmOverloads constructor(
                     TN -> newTn
                     NEXT_VERSION -> null
                     ORIGIN -> origin
-                    GLOBAL_BOOK_FN -> globalBookTn?.featureNumber
-                    CHANGE_COUNT -> (member.readInt64(f) ?: Int64(0)) + 1
+                    GLOBAL_BOOK_FN -> globalBookTn?.featureNumber?.toLong()
+                    CHANGE_COUNT -> (member.readLong(f) ?: 0L) + 1L
                     else -> member.read(f)
                 }
                 val procs = processors[memberName]
@@ -242,13 +245,13 @@ data class Tuple @JvmOverloads constructor(
                         memberValue = proc.processMember(session, collection, f, member, memberValue)
                     }
                 }
-                memberValue = FeatureMemberValues.coerce(memberValue, member.dataType, id, memberName)
+                memberValue = coerce(memberValue, member.dataType, id, memberName)
                 membersBook.put(memberName, memberValue)
             }
 
             // Create the encoder with a custom member encoder.
             val encoder = JbEncoder2(globalBook)
-            encoder.withMemberEncoder { path: Array<Any?>, pathEnd: Int, value: Any? ->
+            encoder.withMemberEncoder { path: Array<Any?>, pathEnd: Int, _: Any? ->
                 // Find the member whose declared path matches the current position and redirect its value into the members-book.
                 members@ for (i in 0 until members.size) {
                     val member = members[i] ?: continue
@@ -281,9 +284,7 @@ data class Tuple @JvmOverloads constructor(
         }
 
         /**
-         * Copy-on-write view of [feature]: every container along the [naksha.model.objects.Member.path]
-         * is replaced by a shallow copy, everything else is shared with [feature].
-         * Member writes into the result therefore never touch [feature].
+         * Copy-on-write view of [NakshaFeature]: every container along the [naksha.model.objects.Member.path] is replaced by a shallow copy, everything else is shared with [NakshaFeature]. Member writes into the result therefore never touch [NakshaFeature].
          */
         fun copyOnWrite(original: NakshaFeature, path: JsonPath): NakshaFeature {
             val root = Platform.copy(original.platformObject(), false) as PlatformMap
@@ -309,16 +310,242 @@ data class Tuple @JvmOverloads constructor(
         private fun isGzipped(bytes: ByteArray): Boolean =
             bytes.size >= 2 && bytes[0] == 0x1F.toByte() && bytes[1] == 0x8B.toByte()
 
+        @Suppress("unused")
         private fun isJbon2(bytes: ByteArray): Boolean =
             bytes.size >= 4 &&
                     bytes[0] == JB2_MAGIC[0] && bytes[1] == JB2_MAGIC[1] &&
                     bytes[2] == JB2_MAGIC[2] && bytes[3] == JB2_MAGIC[3]
 
+        @Suppress("unused")
         private fun isJson(bytes: ByteArray): Boolean {
             val b = bytes[0]
             return b == 0x7B.toByte() || b == 0x5B.toByte() ||
                     b == 0x20.toByte() || b == 0x09.toByte() ||
                     b == 0x0A.toByte() || b == 0x0D.toByte()
+        }
+
+
+        /**
+         * Descend a [NakshaFeature] using the given [path] segments; returns _null_ if any segment is missing.
+         *
+         * @param feature the feature to walk.
+         * @param path the list of path segments (e.g. `["properties", "@ns:com:here:xyz", "updatedAt"]`).
+         * @return the value at the path, or _null_ if the path does not exist.
+         */
+        private fun walkFeature(feature: NakshaFeature, path: List<String>): Any? {
+            var current: Any? = feature
+            for (segment in path) {
+                if (current == null) return null
+                current = when (current) {
+                    is AnyObject -> current.getRaw(segment)
+                    else -> return null
+                }
+            }
+            return current
+        }
+
+        /**
+         * Coerce a raw value to the type expected by the member.
+         *
+         * For [MemberType.SPATIAL], the value is expected to be a [SpGeometry] (or [naksha.geo.SpPoint]) which will be
+         * encoded as TWKB bytes. For other types, the value is expected to already be the raw JSON value.
+         *
+         * @param value the raw value extracted from the feature.
+         * @param type the expected [MemberType].
+         * @param featureId the feature identifier (for logging).
+         * @param memberName the member name (for logging).
+         * @return the coerced value, or _null_ if coercion failed.
+         */
+        private fun coerce(value: Any?, type: MemberType, featureId: String, memberName: String): Any? {
+            if (value == null) return null
+            return when (type) {
+                MemberType.BOOLEAN -> coerceBoolean(value, featureId, memberName)
+                MemberType.INT8 -> coerceInt8(value, featureId, memberName)
+                MemberType.INT16 -> coerceInt16(value, featureId, memberName)
+                MemberType.INT32 -> coerceInt32(value, featureId, memberName)
+                MemberType.INT64 -> coerceInt64(value, featureId, memberName)
+                MemberType.FLOAT32 -> coerceFloat32(value, featureId, memberName)
+                MemberType.FLOAT64 -> coerceFloat64(value, featureId, memberName)
+                MemberType.STRING -> coerceString(value, featureId, memberName)
+                MemberType.BYTE_ARRAY -> coerceByteArray(value, featureId, memberName)
+                MemberType.SPATIAL -> coerceSpatial(value, featureId, memberName)
+                MemberType.TAG_MAP -> coerceTagMap(value, featureId, memberName)
+                MemberType.TAG_LIST -> coerceTagList(value, featureId, memberName)
+                MemberType.TUPLE_NUMBER -> coerceTupleNumber(value, featureId, memberName)
+                else -> {
+                    warnMismatch(featureId, memberName, type.toString(), value)
+                    null
+                }
+            }
+        }
+
+        // -------------------------------------------------------------------------
+        // Individual coercion implementations
+        // -------------------------------------------------------------------------
+
+        private fun coerceBoolean(value: Any, featureId: String, memberName: String): Boolean? = when (value) {
+            is Boolean -> value
+            else -> { warnMismatch(featureId, memberName, "Boolean", value); null }
+        }
+
+        private fun coerceInt8(value: Any, featureId: String, memberName: String): Byte? {
+            val asLong = numberToLongOrNull(value) ?: return null.also { warnMismatch(featureId, memberName, "Byte", value) }
+            if (asLong !in Byte.MIN_VALUE.toLong()..Byte.MAX_VALUE.toLong()) {
+                warnMismatch(featureId, memberName, "Byte", value)
+                return null
+            }
+            return asLong.toByte()
+        }
+
+        private fun coerceInt16(value: Any, featureId: String, memberName: String): Short? {
+            val asLong = numberToLongOrNull(value) ?: return null.also { warnMismatch(featureId, memberName, "Short", value) }
+            if (asLong !in Short.MIN_VALUE.toLong()..Short.MAX_VALUE.toLong()) {
+                warnMismatch(featureId, memberName, "Short", value)
+                return null
+            }
+            return asLong.toShort()
+        }
+
+        private fun coerceInt32(value: Any, featureId: String, memberName: String): Int? {
+            val asLong = numberToLongOrNull(value) ?: return null.also { warnMismatch(featureId, memberName, "Int", value) }
+            if (asLong !in Int.MIN_VALUE.toLong()..Int.MAX_VALUE.toLong()) {
+                warnMismatch(featureId, memberName, "Int", value)
+                return null
+            }
+            return asLong.toInt()
+        }
+
+        // TODO: Needs to become Long!
+        private fun coerceInt64(value: Any, featureId: String, memberName: String): Long? = when (value) {
+            is Int64 -> value.toLong()
+            is Int -> value.toLong()
+            is Long -> value
+            is Short -> value.toLong()
+            is Byte -> value.toLong()
+            is Double -> if (value.isFinite() && value == value.toLong().toDouble()) value.toLong()
+                         else { warnMismatch(featureId, memberName, "Long", value); null }
+            is Float -> if (value.isFinite() && value == value.toLong().toFloat()) value.toLong()
+                        else { warnMismatch(featureId, memberName, "Long", value); null }
+            else -> { warnMismatch(featureId, memberName, "Long", value); null }
+        }
+
+        private fun float32OrNull(value: Any?): Float? = when (value) {
+            is Float -> value
+            is Double -> value.toFloat()
+            is Int -> value.toFloat()
+            is Long -> value.toFloat()
+            is Int64 -> value.toLong().toFloat()
+            is Short -> value.toFloat()
+            is Byte -> value.toFloat()
+            else -> null
+        }
+
+        private fun coerceFloat32(value: Any, featureId: String, memberName: String): Float? {
+            val v = float32OrNull(value)
+            if (v == null) warnMismatch(featureId, memberName, "Float", value)
+            return v
+        }
+
+        private fun float64OrNull(value: Any?): Double? = when (value) {
+            is Double -> value
+            is Float -> value.toDouble()
+            is Int -> value.toDouble()
+            is Long -> value.toDouble()
+            is Int64 -> value.toLong().toDouble()
+            is Short -> value.toDouble()
+            is Byte -> value.toDouble()
+            else -> null
+        }
+
+        private fun coerceFloat64(value: Any, featureId: String, memberName: String): Double? {
+            val v = float64OrNull(value)
+            if (v == null) warnMismatch(featureId, memberName, "Double", value)
+            return v
+        }
+
+        private fun coerceString(value: Any, featureId: String, memberName: String): String? = when (value) {
+            is String -> value
+            is Char, is CharSequence -> value.toString()
+            else -> { warnMismatch(featureId, memberName, "String", value); null }
+        }
+
+        private fun coerceByteArray(value: Any, featureId: String, memberName: String): ByteArray? = when (value) {
+            is ByteArray -> value
+            else -> { warnMismatch(featureId, memberName, "ByteArray", value); null }
+        }
+
+        private fun tupleNumberOrNull(value: Any?): TupleNumber? = when (value) {
+            is TupleNumber -> value
+            is String -> TupleNumber.fromStringOrGuid(value)
+            is ByteArray -> TupleNumber.fromByteArray(value)
+            else -> null
+        }
+
+        /**
+         * Coerce a tuple-number value: an already-materialized [naksha.base.TupleNumber], or a string/byte-array encoding.
+         */
+        private fun coerceTupleNumber(value: Any, featureId: String, memberName: String): TupleNumber? {
+            val v = tupleNumberOrNull(value)
+            if (v == null) warnMismatch(featureId, memberName, "TupleNumber", value)
+            return v
+        }
+
+        /**
+         * Coerce a spatial value to a [SpGeometry].
+         */
+        private fun coerceSpatial(value: Any, featureId: String, memberName: String): SpGeometry? {
+            val geometry = when (value) {
+                is SpGeometry -> value
+                is ByteArray -> fromTWKB(value)
+                else -> value.proxy(SpGeometry::class)
+            }
+            return geometry
+                ?: { warnMismatch(featureId, memberName, "SpGeometry", value); null } as SpGeometry?
+        }
+
+        private fun coerceTagMap(value: Any, featureId: String, memberName: String): TagMap? {
+            val tagMap = value.proxy(TagMap::class)
+            if (tagMap != null) return tagMap
+            warnMismatch(featureId, memberName, "TagMap", value)
+            return null
+            // TODO: Nice Compiler Bug, this code should replace the above, but compiler complains!
+            //       The exact same code works above in `coerceSpatial` ?
+            // return tagMap ?: { warnMismatch(featureId, memberName, "TagMap", value); null } as TagMap?
+        }
+
+        private fun coerceTagList(value: Any, featureId: String, memberName: String): TagList? {
+            val list = when (value) {
+                is TagList -> value
+                is Array<*> -> TagList().apply {
+                    setCapacity(value.size)
+                    for (v in value) if (v is String) addTag(v, true)
+                }
+                else -> value.proxy(TagList::class)
+            }
+            if (list != null) return list
+            warnMismatch(featureId, memberName, "tag_list", value)
+            return null
+        }
+
+        /**
+         * Convert the given number into [Long] or return `null`.
+         * @param value the potential value.
+         * @return either [Long] or `null`.
+         * @since 3.0
+         */
+        private fun numberToLongOrNull(value: Any?): Long? = when (value) {
+            is Byte -> value.toLong()
+            is Short -> value.toLong()
+            is Int -> value.toLong()
+            is Long -> value
+            is Int64 -> value.toLong()
+            is Float -> if (value.isFinite() && value == value.toLong().toFloat()) value.toLong() else null
+            is Double -> if (value.isFinite() && value == value.toLong().toDouble()) value.toLong() else null
+            else -> null
+        }
+
+        private fun warnMismatch(featureId: String, memberName: String, expected: String, value: Any) {
+            logger.warn("Member '$memberName' on feature '$featureId': expected $expected, got ${value::class.simpleName}")
         }
     }
